django_park_tracker_api/requirements.txt
asgiref==3.8.1
bcrypt==4.2.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
